SshKeyGenFragment: use coroutines backed implementation to generate keys
