+++ app/src/main/java/com/zeapo/pwdstore/sshkeygen/SshKeyGenFragment.kt
@@ -10,14 +10,28 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.inputmethod.InputMethodManager
+import android.widget.Toast
+import androidx.core.content.edit
 import androidx.core.content.getSystemService
 import androidx.fragment.app.Fragment
+import androidx.preference.PreferenceManager
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import com.jcraft.jsch.JSch
+import com.jcraft.jsch.KeyPair
 import com.zeapo.pwdstore.R
 import com.zeapo.pwdstore.databinding.FragmentSshKeygenBinding
+import java.io.File
+import java.io.FileOutputStream
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 
-class SshKeyGenFragment : Fragment() {
+class SshKeyGenFragment : Fragment(), CoroutineScope {
 
-    private var keyLength = "4096"
+    override val coroutineContext get() = Dispatchers.IO + Job()
+    private var keyLength = 4096
     private var _binding: FragmentSshKeygenBinding? = null
     private val binding get() = _binding!!
 
@@ -33,7 +47,9 @@ class SshKeyGenFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         with(binding) {
-            generate.setOnClickListener { generate(passphrase.text.toString(), comment.text.toString()) }
+            generate.setOnClickListener {
+                launch { generate(passphrase.text.toString(), comment.text.toString()) }
+            }
             showPassphrase.setOnCheckedChangeListener { _, isChecked: Boolean ->
                 val selection = passphrase.selectionEnd
                 if (isChecked) {
@@ -51,8 +67,8 @@ class SshKeyGenFragment : Fragment() {
             keyLengthGroup.addOnButtonCheckedListener { _, checkedId, isChecked ->
                 if (isChecked) {
                     when (checkedId) {
-                        R.id.key_length_2048 -> keyLength = "2048"
-                        R.id.key_length_4096 -> keyLength = "4096"
+                        R.id.key_length_2048 -> keyLength = 2048
+                        R.id.key_length_4096 -> keyLength = 4096
                     }
                 }
             }
@@ -67,9 +83,45 @@ class SshKeyGenFragment : Fragment() {
     // Invoked when 'Generate' button of SshKeyGenFragment clicked. Generates a
     // private and public key, then replaces the SshKeyGenFragment with a
     // ShowSshKeyFragment which displays the public key.
-    fun generate(passphrase: String, comment: String) {
-        val length = keyLength
-        KeyGenerateTask(requireActivity()).execute(length, passphrase, comment)
+    private suspend fun generate(passphrase: String, comment: String) {
+        withContext(Dispatchers.Main) {
+            binding.generate.text = getString(R.string.ssh_key_gen_generating_progress)
+        }
+        val jsch = JSch()
+        val e = try {
+            val kp = KeyPair.genKeyPair(jsch, KeyPair.RSA, keyLength)
+            var file = File(requireActivity().filesDir, ".ssh_key")
+            var out = FileOutputStream(file, false)
+            if (passphrase.isNotEmpty()) {
+                kp?.writePrivateKey(out, passphrase.toByteArray())
+            } else {
+                kp?.writePrivateKey(out)
+            }
+            file = File(requireActivity().filesDir, ".ssh_key.pub")
+            out = FileOutputStream(file, false)
+            kp?.writePublicKey(out, comment)
+            null
+        } catch (e: Exception) {
+            e.printStackTrace()
+            e
+        }
+        withContext(Dispatchers.Main) {
+            val activity = requireActivity()
+            binding.generate.text = getString(R.string.ssh_keygen_generating_done)
+            if (e == null) {
+                Toast.makeText(activity, "SSH-key generated", Toast.LENGTH_LONG).show()
+                val df = ShowSshKeyFragment()
+                df.show(requireActivity().supportFragmentManager, "public_key")
+                val prefs = PreferenceManager.getDefaultSharedPreferences(activity)
+                prefs.edit { putBoolean("use_generated_key", true) }
+            } else {
+                MaterialAlertDialogBuilder(activity)
+                        .setTitle(activity.getString(R.string.error_generate_ssh_key))
+                        .setMessage(activity.getString(R.string.ssh_key_error_dialog_text) + e.message)
+                        .setPositiveButton(activity.getString(R.string.dialog_ok), null)
+                        .show()
+            }
+        }
         hideKeyboard()
     }
