+++ app/src/main/res/values/strings.xml
@@ -222,6 +222,8 @@
     <string name="ssh_keygen_copy">Copy</string>
     <string name="ssh_keygen_tip">Provide this public key to your Git server.</string>
     <string name="ssh_keygen_show_passphrase">Show passphrase</string>
+    <string name="ssh_key_gen_generating_progress">Generating keys…</string>
+    <string name="ssh_keygen_generating_done">Done!</string>
 
     <!-- Misc -->
     <string name="dialog_ok">OK</string>
